+++ pyengine/Widgets/Label.py
@@ -1,11 +1,12 @@
 from pyengine.Widgets.Widget import Widget
 from pyengine.Utils import Font, Color, Colors
+import pygame
 
 __all__ = ["Label"]
 
 
 class Label(Widget):
-    def __init__(self, position, text, color=Colors.WHITE.value, font=None):
+    def __init__(self, position, text, color=Colors.WHITE.value, font=None, background=None):
         super(Label, self).__init__(position)
         if font is None:
             font = Font()
@@ -14,10 +15,25 @@ class Label(Widget):
             raise TypeError("Font have not a Font type")
         if not isinstance(color, Color):
             raise TypeError("Color have not a Color type")
+        if not isinstance(background, Color) and background is not None:
+            raise TypeError("Background must be a Color")
 
         self.__color = color
         self.__font = font
+        self.__background = background
         self.text = text
+        self.update_render()
+
+    @property
+    def background(self):
+        return self.__background
+
+    @background.setter
+    def background(self, color):
+        if not isinstance(color, Color) and color is not None:
+            raise TypeError("Background must be a Color")
+
+        self.__background = color
         self.update_render()
 
     @property
@@ -54,7 +70,13 @@ class Label(Widget):
         self.update_render()
 
     def update_render(self):
-        self.image = self.font.render().render(self.text, 1, self.color.get())
+        if self.background is None:
+            self.image = self.font.render().render(self.text, 1, self.color.get())
+        else:
+            renderer = self.font.render().render(self.text, 1, self.color.get())
+            self.image = pygame.Surface([renderer.get_rect().width, renderer.get_rect().height])
+            self.image.fill(self.background.get())
+            self.image.blit(renderer, [0, 0])
         self.update_rect()
         if self.parent:
             self.parent.update_render()